// src/types/types.ts
export interface AdvantageTypes {
  id: number;
  title: string;
  text: string;
  image: string;
}

export interface PostsTypes {
  id: number;
  title: string;
  text: string;
  text1: string;
  text2: string;
  text3: string;
  text4: string;
  code1: string;
  code2: string;
}

export interface IntroTypes {
  description: string;
  howItWorks1title: string;
  howItWorks1text: string;
  howItWorks2title: string;
  howItWorks2text: string;
  settingUp: string;
  settingUptext: string;
  settingUptext1: string;
  settingUptext2: string;
  settingUptext3: string;
  settingUptext4: string;
  settingUptext5: string;
}